requirements.txt
asgiref==3.2.10
Django==2.0.4
django-filter==2.0.0
djangorestframework==3.11.0
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1